+++ cypress/integration/a11y.js
@@ -1,5 +1,5 @@
 if (Cypress.env('API') !== 'guillotina') {
-  describe('Accessibility Tests with Axe', () => {
+  describe('Accessibility Tests', () => {
     beforeEach(() => {
       cy.visit('/');
       cy.injectAxe(); // make sure axe is available on the page